implemented the feature to assign students to courses, including the creation of a new template for this functionality. Additionally, I removed the 'enrolled.html' template and made modifications to the 'enrollment.html' template to accommodate the new changes. The commit includes updates to various controllers, entities, repositories, and services to support the new feature and ensure proper functionality across the application.

+++ src/main/java/org/example/springmvcproject/entity/Enrollment.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.LocalDate;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -23,7 +24,7 @@ public class Enrollment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @ManyToOne
     @JoinColumn(name = "student_id")
